requirements.txt
python-telegram-bot==21.6
spotipy==2.24.0
requests==2.32.3
python-dotenv==1.0.1
